Improved FlagsSetterRule so it executes commands right away.
This rule allows a builder-like approach to set its initial flags,
but on host-side tests the commands to set these flags must be
"cached" until the test starts, as ITestDevice (which is used to
call DeviceConfig) is not available right away.

So, initially, the rule was always caching the commands until the
test started, but in reality, that's only needed on host-side tests,
hence this CL abstract that "need" into a new method,
isDeviceConfigAvailable().

Test: atest CtsAdServicesHostTests CtsAdServicesEndToEndTests
Bug: 361771436
Bug: 294423183
Flag: TEST_ONLY

Change-Id: I6d92c60f82a3715b0456d76f10a2b78361236470