devmode: Add flag to guard feature changes
Developer Mode may or may not have implications that need flag guarding
beyond just the shell commands, so create a PhFlag instead of a
DebugFlag. This is because developer mode can potentially put the device
in a bad state so we need the capability to control rollout.

Fixes: 363494565
Test: atest PhFlagsTest
Flag: EXEMPT (bug 337358613)
Change-Id: I4c9ecb94d5945f207f239344e919f35328bf0f08