devmode: Create DevSessionCommand and tests
Introduces a shell command that refreshes developer mode state. This is
done via a new interface DevSessionRefresher, of which the production
implementation will access databases and erase data as well as update
the current state of the dev session.

A follow-up CL will register the command with the adservicesapi shell
command factory, plus some end-to-end-tests.

Flag: EXEMPT (bug 337358613)
Fixes: 363511595
Test: atest DevSessionCommandTest
Change-Id: I26235a56530c39653ef9ff15a535173e3f990022