Set PRECOMPUTED_CLASSIFIER as default in tests.
Classifier lib is known to have issue on 32bit devices. To fix this, we
are rolling out a flag that will set PRECOMPUTED_CLASSIFIER as the
default classifier.
In this CL, we are setting PRECOMPUTED_CLASSIFIER as the default
classifer to avoid having issues on 32bit devices.

Test: atest FledgeCtsDebuggableTest
Flag: TEST_ONLY
Bug: 357590057
Change-Id: I775df061668ea627c73cfc12b94c7133eef247f3