Set and use flags instead of hardcoding in mocks
This CL modifies the FetchCustomAudienceImplTest. In this test class we
are mocking CustomAudienceServiceFilter and we are expecting certain
arguments in the mocked methods. These methods are mocked with hard
coded values in the argument. However in the service code, these values
are picked from the flags, resulting in mismatch of expected vs actual
argument because these flag values might have been set to a different
value in some other test.

This CL makes sure that the expected flag values are properly set in the
tests.

Test: FetchCustomAudienceImplTest
Bug: 367520509
Flag: TEST_ONLY
Change-Id: Ia6a2d3a0c28ad131f9fb3eb9ce44e8827052edd9