Set correct flag in ReportAndRegisterFallbackTest
We should not rely on the default value of the Flag. If we expect to use
a flag, we should set the flag in the tests. This is even more important
in DebugFlags because these flags are SystemProperties and the default
method doesn't always return the default value.

Fix: 368213636
Test: atest ReportAndRegisterEventFallbackImplTest
Flag: TEST_ONLY
Change-Id: Ic7567ca418b9f9f162169e59bd9e4a4b9d776fe9