Set debug flag value correctly in tests
We were expecting the default method in instances of Flags class to return the
default value. However this is not true in case of DebugFlags because
these values are picked up from the SystemProperties.
This sometimes lead to flakiness in tests.
This CL makes sure that we are setting the correct value of flag in the
tests.

Test: atest AdServicesServiceCoreTopicsUnitTests
Bug: 368023681
Flag: TEST_ONLY
Change-Id: Ic78013c2d07d62d3045b9860da44820cb0e18587