Merged jobMocker into mocker.
The 'mocker' reference should follow the "Highlander Rule" ("There
Can be Only One!"), so developers can type mocker. and let the IDEs
auto-complete the options.

Prior to this CL, that was not possible because the mocker references
were final; this CL "splits" the base superclasses (like
AdServicesExtendedMockitoTestCase) in 2 classes: the "real" one and a
"mocker-less" superclass. That way, subclasses (like
AdServicesJobServiceTestCase) can provide their own custom "mocker".

Fixes: 369656942
Test: atest AdServicesTestUtilityTests AdServicesSharedLibrariesUnitTests
Test: atest AdServicesServiceCoreMeasurementUnitTests --test-filter '.*JobServiceTest.*'
Flag: TEST_ONLY

Change-Id: I8c2ac9a8d1d7cd26c04fa9252f7fca4add681999