devmode: Throw ISE if storage not initialized
Make initializing the DevSessionStorage proto an explit operation,
rather than relying on DevSessionState#UNKNOWN to represent it. When
invoking the first get() from storage, check if is_initialized is set
and update to IN_PROD if not.

Finally, if there is an issue with the storage we can distinguish
between an UKNOWN state and an uninitialized state. This will become
especially useful if/when we add the expiry field back.

Fixes: 371711645
Test: atest AdServicesServiceCoreProtectedAudienceUnitTests
Flag: EXEMPT (bug 337358613)
Change-Id: I2f5b0bf347421c11f35e43f51632618aa75a0470