devmode: Throw ISE if DevSessionState is UNKNOWN
Make sure to never allow PPAPI calls during an UNKNOWN state. When the
flag is disabled IN_PROD is always the value, so this change will not
put traffic at risk outside a dev session. If we enter an unrecoverable
state then we need to inform the user how to exit, so create a new set
of status codes and propagate that back to clients.

Test: atest AdServicesServiceCoreUnitTests && \
  AdServicesServiceCoreProtectedAudienceUnitTests
Flag: EXEMPT (bug 337358613)
Fixes: 371711645
Fixes: 372082802
Change-Id: Iffeeb2d453610fec73835066775a0ae2343c0437