devmode: Add test for caller SE rejection & fix race condition
Ensure that all non-debuggable callers are rejected with a
SecurityException during dev mode. This smoke test ensures this
behaviour as the CTS package is non-debuggable. Note that these tests
will need to migrate to a new, dedicated non-debuggable unittest package in the future.

Also fixes a race condition in the DevSessionControllerImpl where each
chained operation was executing immediately after the previous one
started.

Bug: 363512743
Bug: 363480549
Bug: 372092931
Fixes: 372214323
Fixes: 363511599
Test: atest CtsAdServicesDeviceTestCases && \
  CtdAdServicesDebuggableDeviceTestCases
Test: atest AdServicesServiceCoreUnitTests && \
  AdServicesServiceCoreProtectedAudienceUnitTests
Flag: EXEMPT (bug 337358613)
Change-Id: I23602b032e307c7cd00f880a86b0e19370552139