Fix EpochJobServiceTest
We have some tests which schedules the job twice and then checks the
argument used using ArgumentCaptor. This test relies on the order of
execution of this two methods. If we use an executor - then we cannot guarantee the order and this is making the tests flaky.

This CL changes the tests and makes sure we don't rely on the order of
the execution of the method in the tests.

Test: atest EpochJobServiceTest
Fix: 367831288
Flag: EXEMPT (bug 337358613)
Change-Id: I97a4454ef42f6dd5c59cb145e38803dd6000bff3