Refactor SandboxedAppSetIdManagerTest to use FlagsSetterRule.
As apart of an audit to ensure all AdServices CTS tests are setting
flags through the FlagsSetterRule, there is one class that needs to
be fixed.

This is needed for the validity of flag ramp testing runs so
that all flags that are set for the test case are reset to their
previous state at the end.

Also added an interface for setting common flags for
SandboxedAppSetIdManagerTest.

Bug: 356893692
Test: atest SandboxedAdIdManagerTest SandboxedAppSetIdManagerTest
Flag: TEST_ONLY
Change-Id: Iacc752a3749e36f74ca8675c6d4400107353447a