Remove flag sync commands from CTS AndroidTest files.
As part of the audit to ensure validity of flag ramp CTS test runs,
it would be best for tests to not enable flag sync to avoid overriding
the desired launch flags that were set on the device initially.

There are few tests that still seem to be doing it. Considering we have
introduced FlagsPreparerClassRule, which automatically disables flag
sync, this extra setup in the AndroidTest file shouldn't be needed.

Also added @SetSdkSandboxStateEnabled for appropriate tests.

Will monitor for test flakiness.

Bug: 356893692
Bug: 361771436
Test: atest CtsSandboxedAdIdManagerTests CtsSandboxedMeasurementManager CtsSandboxedAppSetIdManagerTests
Flag: TEST_ONLY
Change-Id: I9936e66ae61c72cd0b01c39179331393cb457281