IOException in Http failues in SCAUpdate telemetry
This CL catches and logs IO Exception in Schedule Custom Audience
telemetry. IO Exception is thrown by the AdServicesHttp client when the
size of the payload in the response exceeds the max limit.
This telemetry will be useful to determine if the MMP are sending
payloads large payloads.

Test: Atest ScheduleUpdatesHandler
Bug: 375485307
Flag: EXEMPT (bug 337358613)
Change-Id: I0c7c3e8510975960df3c72ddb6799b145b63b9d3