Remove Foreground Activity for some PA Shell Cmd CTS Tests
These tests are currently ~100% failing on a few targets and in my local. The error implies they timeout for waiting for broadcast android.adservices.utils.SimpleActivity.RESULT.

It seems the Rb Shell Command related CTS tests are conflict with the foreground activity. Moreover, to use foreground activity is not needed for such tests.

Remove the foreground activity usage for these tests to mitigate the issue.

Flag: TEST_ONLY
Bug: 377722576
Test: atest CtsAdServicesDebuggableDeviceTestCases --rerun-until-failure 2
Change-Id: I201be7d16f16fc68ca5e17272abb451d715faed9