Preparing AdServicesSharedLibrariesUnitTestsRavenwood
... for upcoming ravenwood change.

Ravenwood will start taking the test package name from the build file,
and stop taking it from the config.

(Long term, we'll read it from the manifest, but for now we'll get it
from the build file.)

That change will happen in AOSP. We'll update DeviceSideTestCase before
that, so that AdServicesSharedLibrariesUnitTestsRavenwood will
keep passing even after the change is merged.

Flag: EXEMPT host test change only
Bug: 377765941
Test: atest AdServicesSharedLibrariesUnitTestsRavenwood
    With and without the upcoming ravenwood change.

Change-Id: I01e4cc67f97045ffec8b749d262aedbcada4c983