add new errorlog method to log the root exception
this is a change to add a method to log the stack trace from the farthest avilable cause. current log method log the first exception of the stacktrace, it won't have useful information for future calls, which will always wrap the inner exception to ExecutionException. This method can get information smarter.

Bug: 370523314
Test: atest com.android.adservices.shared.errorlogging.AbstractAdServicesErrorLoggerTest
Flag: EXEMPT (bug 337358613)
Change-Id: Id58f8c8f79ba840e9d44f206e00e694a91c79e26