Remove one ReportEventRequestCts test
Currently the valid values for the destination is 0-8, this test is
passing because the test is expecting that an error should be thrown
when the destination value is 9.
However in future, we might be adding another destination which will
make valid destination values to be 0-16. And this test will start
failing. Because this is a cts tests, we have to make sure that future
versions of the code also work with this test (new code old test).

Test: Atest CtsAdServicesDeviceTestCases:ReportEventRequestTest
Bug: 381205592
Flag: TEST_ONLY
PS_FLAG: TEST_ONLY

Change-Id: I4c2cc8766e1430e2bcdf5d7d4724868b2b7dce83