Removed custom AtomicFileDatastore subclass.
AtomicFileDatastore was initially created in the adservices project,
but eventually moved to the shared library, so it could be used by
OnDevicePersonalization. But to minimize changes, the shared class
was non-final and extended by the "legacy" one - this change gets
rid of the legacy one, as its usage has been replaced by a factory
(LegacyAtomicFileDatastoreFactory, which returns instances of the
shared class).

Test: atest AdServicesSharedLibrariesUnitTests:AtomicFileDatastoreTest AdServicesTestUtilityTests AdServicesUIUnitTests AdServicesServiceCoreUnitTests:LegacyAtomicFileDatastoreFactoryTest
Bug: 372051957
Flag: EXEMPT (bug 337358613)
BYPASS_PS_FLAG_CHECK: infra-related change (not backed up by a flag)

Change-Id: I16695a9b2ad1eefd483834bc6022791a6ac145a3