Update storage tests to support HSUM
The CL also fixes a bug where we were not listening for package
installation broadcast for all users.

The broadcast receiver is an optimization where we create the per-sdk
storage immediately when a package is installed. It doesn't impact the
overall picture since we have a fallback during loadSdk call where we
create per-sdk storage directories if they are missing.

Bug: 372475678
Test: lunch <some-hsum> && atest SdkSandboxStorageHostTest
Flag: EXEMPT b/337358613
BYPASS_PS_FLAG_CHECK: b/391335182

Change-Id: Ie4da35db156c2b836af9e2b016e3f018bb93be26