Minor changes on some FlagSetterRule methods.
The setFlag(String, String...) was renamed to setArrayFlag(...); not
only it would be consistent with setArrayFlagWithExplicitSeparator(),
but it also makes room for a new, generic setFlag(String, Object).

As such, it also re-added a setString(String, String) method (as that
method was not needed when the String... version was added).

Test: atest AdServicesSharedLibrariesUnitTests --test-filter=".*FlagsSetterRule.*"
Test: atest AdServicesTestUtilityTests --test-filter=".*FlagsSetterRule.*"
Bug: 384798806
Flag: TEST_ONLY
BYPASS_PS_FLAG_CHECK: test-only change

Change-Id: Id477ad49a1d2d21705cd5fd5edf12aea15142108