Refactor CachedFlagsTest
This is to prepare to migrate this class to use the Process Stable Flags framework.

- Use Rb Testing framework
- Create helper methods to avoid duplicates
- Clean up the test logic a bit.

Note the flag implementation does NOT follow the convention. mEnableSessionStableKillSwitches is a KillSwitch, which should disable the feature when it's enabled. This CL doesn't modify the production code.

Flag: TEST_ONLY
BYPASS_PS_FLAG_CHECK: test only.
Bug: 385741751
Test: atest CachedFlagsTest

Change-Id: I640cafaf2d889d727662e438a852f282d0c2edce