Fix configurations proto structure.
As per the final design, configurations of each type will be served individually (example, with their own mdd setup). Therefore the Configuration will not be repeated.

The binary sent from the server will be of type ConfigurationPayload.

Will send a followup cl to update the proto schema change in the server.

Bug: 378678172
BYPASS_PS_FLAG_CHECK: This proto is not yet referenced in code.
Flag: EXEMPT (bug 337358613)
Change-Id: I56d1c70576bdf8ae5a940b771150a4a51caeb8d4