Part 2: Implementation of device level sampling
The logic for device selection is similar to Gmscore.
The device id/selection id is calculated once and then store in
datastore. Subsequent calls use this device id to determine whether to
log the event or not.

Hash function uses device id, group id and period number to generate the
id and then make the decision for logging.

Follow-up cl with integrate this with the
MetricLogger/AbstractMetricLogger.

Test: atest AdServicesSharedLibrariesUnitTests:com.android.adservices.shared.metriclogger.logsampler.deviceselection.DeviceSelectionLogicTest
Flag: EXEMPT (bug 337358613  )
BYPASS_PS_FLAG_CHECK: not used in prod yet
Bug: 385201138

Change-Id: I054767b3915423501239693ece030a894a31944e