Mock ErrorLogUtil instead of spying.
Mocking ErrorLogUtil will genereally require ErrorLogUtil to be mocked
in case it's used in source code; this is done automatically via the
AdServicesLoggingUsageRule. In case the rule is skipped, spying is
causing SecurityException issues due to flag checks that need to be
accounted for in the source code in case it's not mocked properly.
Mocking instead of spying should hopefully make the error easier to
understand.

Bug: 384574337
Test: presubmit
FLAG: TEST_ONLY
BYPASS_PS_FLAG_CHECK: test-only change

Change-Id: I01f820f7e7b199b066642123aad8b2446893b768