Implement versioned signal update download
This change adds a new download path to UpdatesDownloader that both
requests and expects in its response the schema version number for the
downloaded updates.

This returned update schema version is then bundled with the update JSON
in a new SignalUpdates object and passed to the update processing
orchestrator.

Currently, the version is otherwise unused; a follow-up change will
modify the update processing flow to handle version compatibility in
processing.

Bug: 345461789
Test: atest UpdatesDownloaderTest
Test: atest UpdateProcessingOrchestratorTest
Test: atest UpdateSignalsOrchestratorTest
Test: atest ForcedEncodingE2ETest SignalsEncodingE2ETest
Test: atest SignalsIntakeE2ETest TriggerEncodingCommandE2ETest
Flag: EXEMPT (bug 337358613)
PS_FLAG: ProtectedSignals__enable_prioritized_eviction

Change-Id: Id642022bc75a0acf9136282b2d968c1d4ac88856