Add failure callback for AttributionFallbackJobService
Exceptions that occur within the work submitted to the background thread
via the submit method go uncaught, resulting in silent failure. This is
different from the behavior of work submitted via the execute method.
For JobServices this means the reference to the params object is lost
and jobFinished is never called, resulting in abandoned jobs. To
mitigate this, we add failure callbacks to ensure jobFinished is
still called.

Bug: 406789553
Bug: 401056315
Flag: EXEMPT (bug fix)
Test: atest AdServicesServiceCoreMeasurementUnitTests:com.android.adservices.service.measurement.attribution.AttributionFallbackJobServiceTest
BYPASS_PS_FLAG_CHECK: bug fix

Change-Id: I6e492f43915c96fd6cf8c07f9b5b2cb70f4067f7