Use a different lock for mRunningInstrumentations
A deadlock could occur when ActivityManagerService calls
ActivityInterceptorCallback.onInterceptActivityLaunch() because
mRunningInstrumentations is accessed which is locked by the common lock
in SdkSandboxManagerService. If SdkSandboxManagerService then calls into
ActivityManagerService with this common lock, there would be a deadlock.

Use a separate lock for mRunningInstrumentations to avoid this problem.

Bug: 404187427
Bug: 400376755
Test: atest
Flag: EXEMPT b/337358613
PS_FLAG: fix_deadlock_bug_398296192

Change-Id: I97bed9c187f246c1a826ced57f04143f1f944fda