Partially revert proto flag changes in "Refractor Argon Config Delivery Flags"
This change reverts the Proto flag related changes in ag/33442721.

The cause of the regression is the use of Base64 conversion for the proto flags introduced in ag/33442721.

The following changes in ag/33442721 are not reverted in this change, as cls submitted post ag/33442721 depend on them and they are confirmed to be not the culprit for the regression.
  * Removal of enable_enrollment_config_v3_db
  * Rename of use_configs_manager_to_query_enrollment to use_argon_config_manager_to_query_enrollment
  * Addition of use_argon_config_manager_to_query_enrollment

This partially reverts commit 1fa1ec9106326e3140bbb339fd80df07d5cda795.

Bug: 417597867
Reason for revert: Causing regression in showmap_pss_bytes_imperceptible_com.google.android.adservices.api
BYPASS_PS_FLAG_CHECK: Revert change
Flag: EXEMPT (bug 337358613)
BYPASS_FLAG_NAME_CHECK: Only removing a flag

Change-Id: I8299981f2c0002d1685b0097a4bcfeffd9392fc9