Fix uninstalling appsetidsdk after the test runs
SuiteApkInstaller is not able to uninstall the AppSetIdSdk.apk because
the test app that requires it is still installed at that point. This
means the apk would stay on the test device after the test ends.

This preparer, during tear down, runs after SuiteApkInstaller and will
finally uninstall the provider, leaving the device in a clean state.

Bug: 432324494
Test: atest CtsSandboxedAppSetIdManagerTests
Flag: TEST_ONLY
BYPASS_PS_FLAG_CHECK: TEST_ONLY

Change-Id: I15eac227d9fd1c49dfa323a858c8c5195b479cac