Baseline NewApi failures in AdServices module
With the upcoming command line tools update in ag/34989485,
we are baselining the existing lint issues.

As the module is not actively maintained, it is better to baseline these
issues than attempt to fix them at this point in time. The issues are
blocking the upgrade.

Test: m lint-check with topic "lint-upgrade"
Bug: 452688941
Bug: 441794572

Flag: NONE update lint baseline
BYPASS_PS_FLAG_CHECK: update lint baseline

Change-Id: I24847aa970d4b33d871d42ce523024ea737cf378